BUG: Remove bad assert leading to debug failures
The assert doesn't make sense for richcompare, since it has no
notion of forward.
It is also problematic in the other cases, because e.g. complex
remainder defers (because complex remainder is undefined).  So
`complex % bool` will ask bool, which then will try to defer to
`complex` (even though it is not a forward op).

That is correct, since both should defer to tell Python that the
operation is undefined.