ENH: Check floating point error flags for all casts
To achieve this, I first need to propagate the ArrayMethod information
that FPE flags _may_ be given for a specific cast.
This requires quite a bit of changes (which is arguably the more annoying
part of the PR/commit).
Note that for some "legacy" casts, we do not modify this and assume that
the only interesting information is `needs_api`.