MAINT: Ignore warning rather than forcing output dtype
Forcing the output dtype does not work here, since it actually
can be integral (not that this is usually a good idea).
In practice, the change we are doing here is forcing 64bit (or
32bit depending on platform) of precision for the calculation.

This means that the change will only ever increase precision mildly
compared to the current situation.