ENH: Try to be more forgiving in abstract dtype promotion
This follows the tests (and actually goes hand in hand with them).
There are still some apparent issues here though, I suspect (but
am not sure), the that the legacy promotion may need to kick in
more agressively when errors occur.

Also, surprisingly this fixes things that maybe should fail in
legacy promotion, and I am not yet sure why...