ENH: Always fill object fields with None rather than NULL
The intention here is to transition (or achieve?) a future where
NumPy always fills object fields with None rather than NULL.

Filling with NULL is nice (infinitely fast in some cases and easy),
but it also adds the trap of expecting NULLs everywhere.
Most importantly, cython fails to do so (currently) so this is a
bug magnet.

At this point, ideally any fully created array is guaranteed to be
initialized with proper Python objects.  Downstream may break that,
so we should not allow it, but this does also add an assert to
many ufuncs to flush out further potential issues.

NULL initialization is (as of now) still relevant for new buffers.
That means that core functionality like `setitem`, or coping into
a buffer should still expect a NULL initialized value.