MAINT, TST: Shorten wheel test output on travis builds
Travis has a limit on log length that is causing test failures, so
decrease the verbosity of the test output. Travis has been on the edge
for a while and less drastic fixes haven't sufficed.

The choice here is to use the default pytest verbosity as "-v" is hard
to read and all greater verbosities are very long.