DEP: Deprecate out-of-bound Python integer conversions
Any conversion from a Python integer (or subclass) that is stored
into a NumPy dtype but does not fit should raise an error in the future.
Note, that casts between NumPy types (or assignments of them) are
explicitly not affected by this.

There are certain use-cases for allowing such casts, even if
technically undefined behavior in C.  They just work out well
in practice in many cases since e.g. -1 is all 1's in binary
represenation (twos complement repr).