MAINT: Structured MA fill value workaround by adding array cast
This wraps the fill value into an array, the default fill value for all
ointegers is 99999 which doesn't work for many integer dtypes.

Note that this might still subtle change the behavior in other
code paths where we cannot avoid this.  Plus, the deprecationwarning
may show up (and in fact be a "in the future will use the default
fill value" warning).