BUG: Fix use and errorchecking of ObjectType use
This should be replaced really, it is pretty bad API use, and doesn't
work well (up to being incorrect probably).

But working on other things (trying to make promotion strict and thus
saner), I realized that the use was always wrong: we cannot pass 0
since 0 means `bool`, what was always meant was passing no-type.

So fixing this, and adding the error check everywhere.  Checking
for `PyErr_Occurred()` may have been necessary at some point, but
is not anymore.