DEP: Next step in scalar type alias deprecations/futurewarnings
Finalizes the scalar type alias deprecations making them an error.
However, at the same time adds a `FutureWarning` that new aliases
will be introduced in the future.
(They would eventually be preferred over the `str_`, etc. version.)

It may make sense, that this FutureWarning is already propelled soon
since it interacts with things such as changing the representation of
strings to `np.str_("")` if the preferred alias becomes `np.str`.

It also introduces a new deprecation to remove the 0 sized bit-aliases
and the bitsize `bool8` alias.  (Unfortunately, these are here still allowed
as part of the `np.sctypeDict`).